Inital translations
I tried hard not to add literal translations but to communicate the idea, maybe now the English version is a bit out of sync since the concept of the site has shifted since I first wrote the english pages. We may need to revisit the english pages as well.

The work is pretty straightforward, hugo enables translations by prefixing the language name to the file extensions. For example `contents/index.md` can be translated to spanish by creating a new file named `contents/index.es.md`.

There is still somre repeated code on the index page since it's not only content but some styles and the newsletter form, I think for now this works but if we find ourselves changing a lot the index page then we should move these resources to a more generic place.